refactor: Split received and sent messages into separate components
